QS: Enable ProfileTile (2/2)
Enables the ProfileTile to control the Profile currently in use.

This was partially implemented but had various parts missing, due to QuickSettings being overhauled I was unable to cherry-pick the original CM addition of the ProfileTile as it was originally authored under the old QuickSettings implementation.

Compiled and tested with no issues, changes profile correctly.

Change-Id: I63cd5fcb9e1b9cebea01c4ad122050e479334539