FW: Fix Notification Shortcuts
All credits to SlimRoms for their awesome work on the Notifications shortcuts, this commit corrects the PAC code to the SlimRoms 4.3 implementation.

Unable to cherry-pick a commit to correct our code issues but the original commit retains authorship and all credits for this code are solely attributed to the SlimRoms team, I have only corrected the PAC code port of this feature.

Change-Id: Ib74118e85c31778a24fb52ffec68cfef6f5a9d2a