Add checkbox in recents settings to make app shortcuts in recents optional
Since some time there is a shortcut bar in recents for some selected apps.
I don't need it and it clutters the interface. Maybe somebody else likes
the shortcuts so I thought I'd just make it optional.
This is only the Settings part of it. Other part is in frameworks/base
and needs this patch.

Change-Id: I0f365616e048292e43552bcf84bf24798d6432b9