Fix mismatch theme elements
Most themers use CM as the base and refuse to theme icons for other ROMs.
This leaves us in the cold and left with mismatch themes or forced to track
these guys down and beg for things to be themed......is a pain in the ass

This fixes that problem by way of tinting existing icons using the material_deep_teal_500
hex color. Once they change the hex (material_deep_teal_500) it will tint our elements as well.

Change-Id: I1775142c894cb6f390ae24299eec47f04fa6cbb1
Signed-off-by: Lokesh Chamane <lokesh.c703@gmail.com>